add antigen, reorder script install to install zsh

+++ scripts/dotnet_install.sh
@@ -1,6 +1,13 @@
-# get lastet version of the dotnet install script
+# get dotnet install script if not installed
+
+if ! [[ -f ~/.scripts/dotnet-install.sh ]];
+then
+
 wget https://dot.net/v1/dotnet-install.sh -O ~/.dotfiles/scripts/dotnet-install.sh
 
 chmod +x ~/.dotfiles/scripts/dotnet-install.sh
+
+fi
+
 
 ~/.dotfiles/scripts/dotnet-install.sh -c Current
